refactor(internal): split utils between Jessie and extra-Jessie
