+++ packages/internal/src/index.js
@@ -3,3 +3,4 @@
 export * from './config.js';
 export * from './debug.js';
 export * from './utils.js';
+export * from './method-tools.js';
