+++ packages/internal/src/utils.js
@@ -1,19 +1,12 @@
 // @ts-check
+// @jessie-check
 import { E } from '@endo/eventual-send';
 import { deeplyFulfilled, isObject } from '@endo/marshal';
 import { isPromise } from '@endo/promise-kit';
-import { asyncGenerate } from 'jessie.js';
+import { asyncGenerate, makeSet } from 'jessie.js';
 
-/** @typedef {import('@endo/marshal/src/types').Remotable} Remotable */
-
-const {
-  getPrototypeOf,
-  create,
-  entries,
-  fromEntries,
-  getOwnPropertyDescriptors,
-} = Object;
-const { ownKeys, apply } = Reflect;
+const { entries, fromEntries } = Object;
+const { ownKeys } = Reflect;
 
 const { details: X, quote: q, Fail } = assert;
 
@@ -32,7 +25,7 @@ export const fromUniqueEntries = allEntries => {
   if (ownKeys(result).length === entriesArray.length) {
     return result;
   }
-  const names = new Set();
+  const names = makeSet();
   for (const [name, _] of entriesArray) {
     if (names.has(name)) {
       Fail`collision on property name ${q(name)}: ${entriesArray}`;
@@ -95,7 +88,7 @@ harden(objectMap);
  * @param {Array<string | symbol>} rightNames
  */
 export const listDifference = (leftNames, rightNames) => {
-  const rightSet = new Set(rightNames);
+  const rightSet = makeSet(rightNames);
   return leftNames.filter(name => !rightSet.has(name));
 };
 harden(listDifference);
@@ -150,103 +143,6 @@ export const applyLabelingError = (func, args, label = undefined) => {
   }
 };
 harden(applyLabelingError);
-
-/**
- * Prioritize symbols as earlier than strings.
- *
- * @param {string|symbol} a
- * @param {string|symbol} b
- * @returns {-1 | 0 | 1}
- */
-const compareStringified = (a, b) => {
-  if (typeof a === typeof b) {
-    const left = String(a);
-    const right = String(b);
-    // eslint-disable-next-line no-nested-ternary
-    return left < right ? -1 : left > right ? 1 : 0;
-  }
-  if (typeof a === 'symbol') {
-    assert(typeof b === 'string');
-    return -1;
-  }
-  assert(typeof a === 'string');
-  assert(typeof b === 'symbol');
-  return 1;
-};
-
-/**
- * TODO Consolidate with the `getMethodNames` in `@endo/eventual-send`
- *
- * @param {any} val
- * @returns {(string|symbol)[]}
- */
-export const getMethodNames = val => {
-  let layer = val;
-  const names = new Set(); // Set to deduplicate
-  while (layer !== null && layer !== Object.prototype) {
-    // be tolerant of non-objects
-    const descs = getOwnPropertyDescriptors(layer);
-    for (const name of ownKeys(descs)) {
-      // In case a method is overridden by a non-method,
-      // test `val[name]` rather than `layer[name]`
-      if (typeof val[name] === 'function') {
-        names.add(name);
-      }
-    }
-    if (!isObject(val)) {
-      break;
-    }
-    layer = getPrototypeOf(layer);
-  }
-  return harden([...names].sort(compareStringified));
-};
-harden(getMethodNames);
-
-/**
- * TODO This function exists only to ease the
- * https://github.com/Agoric/agoric-sdk/pull/5970 transition, from all methods
- * being own properties to methods being inherited from a common prototype.
- * This transition breaks two patterns used in prior code: autobinding,
- * and enumerating methods by enumerating own properties. For both, the
- * preferred repairs are
- *    * autobinding: Replace, for example,
- *      `foo(obj.method)` with `foo(arg => `obj.method(arg))`. IOW, stop relying
- *      on expressions like `obj.method` to extract a method still bound to the
- *      state of `obj` because, for virtual and durable objects,
- *      they no longer will after #5970.
- *    * method enumeration: Replace, for example
- *      `Reflect.ownKeys(obj)` with `getMethodNames(obj)`.
- *
- * Once all problematic cases have been converted in this manner, this
- * `bindAllMethods` hack can and TODO should be deleted. However, we currently
- * have no reliable static way to track down and fix all autobinding sites.
- * For those objects that have not yet been fully repaired by the above two
- * techniques, `bindAllMethods` creates an object that acts much like the
- * pre-#5970 objects, with all their methods as instance-bound own properties.
- * It does this by making a new object inheriting from `obj` where the new
- * object has bound own methods overridding all the methods it would have
- * inherited from `obj`.
- *
- * @param {Remotable} obj
- * @returns {Remotable}
- */
-export const bindAllMethods = obj =>
-  harden(
-    create(
-      obj,
-      fromEntries(
-        getMethodNames(obj).map(name => [
-          name,
-          {
-            value: (/** @type {unknown[]} */ ...args) =>
-              apply(obj[name], obj, args),
-            enumerable: true,
-          },
-        ]),
-      ),
-    ),
-  );
-harden(bindAllMethods);
 
 /**
  * @template T
@@ -304,7 +200,7 @@ export const makeMeasureSeconds = currentTimeMillisec => {
  * @param {string} [message]
  */
 export const makeAggregateError = (errors, message) => {
-  const err = new Error(message);
+  const err = Error(message);
   Object.defineProperties(err, {
     name: {
       value: 'AggregateError',
